+++ resources/views/admin/providers/index.blade.php
@@ -27,7 +27,9 @@
                     <div class="card-body">
                         <div class="d-flex justify-content-between">
                             <h4 class="card-title">Proveedor
-                                <a href="{{route('providers.create')}}" class="btn btn-primary">Crear Nuevo</a>
+                                @can('providers.create')
+                                    <a href="{{route('providers.create')}}" class="btn btn-primary">Crear Nuevo</a>
+                                @endcan
                             </h4>
                             <div class="btn-group">
                                 <a data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
@@ -63,17 +65,16 @@
                                         <td>{{$provider->address}}</td>
                                         <td>{{$provider->phone}}</td>
                                         <td style="width: 50px;">
+                                        @can('providers.destroy','providers.edit')
                                             {!! Form::open(['route'=>['providers.destroy',$provider], 'method'=>'DELETE']) !!}
-
                                             <a class="jsgrid-button jsgrid-edit-button" href="{{route('providers.edit', $provider)}}" title="Editar">
                                                 <i class="far fa-edit"></i>
                                             </a>
-
                                             <button class="jsgrid-button jsgrid-delete-button unstyled-button" type="submit" title="Eliminar">
                                                 <i class="far fa-trash-alt"></i>
                                             </button>
-
                                             {!! Form::close() !!}
+                                        @endcan
                                         </td>
                                     </tr>
                                     @endforeach
