+++ resources/views/admin/role/index.blade.php
@@ -16,7 +16,7 @@
             </h3>
             <nav aria-label="breadcrumb">
                 <ol class="breadcrumb">
-                    <li class="breadcrumb-item"><a href="#">Panel administrador</a></li>
+                    <li class="breadcrumb-item"><a href="{{ route('home') }}">Panel administrador</a></li>
                     <li class="breadcrumb-item active" aria-current="page">Roles</li>
                 </ol>
             </nav>
@@ -25,23 +25,12 @@
             <div class="col-lg-12 grid-margin stretch-card">
                 <div class="card">
                     <div class="card-body">
-                        
+
                         <div class="d-flex justify-content-between">
-                            <h4 class="card-title">Roles 
+                            <h4 class="card-title">Roles
                                 <a href="{{route('roles.create')}}" class="btn btn-primary">Crear Nuevo</a>
                             </h4>
-                            {{--  <i class="fas fa-ellipsis-v"></i>  --}}
-                            <div class="btn-group">
-                                <a data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                                    <i class="fas fa-ellipsis-v"></i>
-                                </a>
-                                <div class="dropdown-menu dropdown-menu-right">
-                                  {{--  <button class="dropdown-item" type="button">Another action</button>
-                                  <button class="dropdown-item" type="button">Something else here</button>  --}}
-                                </div>
-                              </div>
                         </div>
-    
                         <div class="table-responsive">
                             <table id="order-listing" class="table">
                                 <thead>
@@ -63,17 +52,16 @@
                                             {{ $rol->guard_name }}
                                         </td>
                                         <td style="width: 50px;">
-                                            {!! Form::open(['route'=>['roles.destroy',$rol], 'method'=>'DELETE']) !!}
-    
-                                            <a class="jsgrid-button jsgrid-edit-button" href="{{route('roles.edit', $rol)}}" title="Editar">
-                                                <i class="far fa-edit"></i>
-                                            </a>
-                                            
-                                            <button class="jsgrid-button jsgrid-delete-button unstyled-button" type="submit" title="Eliminar">
-                                                <i class="far fa-trash-alt"></i>
-                                            </button>
-    
-                                            {!! Form::close() !!}
+                                            @can('roles.destroy','roles.edit')
+                                                {!! Form::open(['route'=>['roles.destroy',$rol], 'method'=>'DELETE']) !!}
+                                                <a class="jsgrid-button jsgrid-edit-button" href="{{route('roles.edit', $rol)}}" title="Editar">
+                                                    <i class="far fa-edit"></i>
+                                                </a>
+                                                <button class="jsgrid-button jsgrid-delete-button unstyled-button" type="submit" title="Eliminar">
+                                                    <i class="far fa-trash-alt"></i>
+                                                </button>
+                                                {!! Form::close() !!}
+                                            @endcan
                                         </td>
                                     </tr>
                                     @endforeach
